refactor: split ThumbnailCache and ScrollWheelRedirector out of ThumbnailStripView
Extract the NSCache-backed thumbnail cache (and its sole consumer
AsyncThumbnailImage) into ThumbnailCache.swift, and the vertical->horizontal
scroll-wheel redirector into ScrollWheelRedirector.swift, so
ThumbnailStripView.swift holds only the SwiftUI view concern. Promotes the
previously private ThumbnailCache to internal for cross-file access and drops
four leftover debug print statements from the scroll monitor.
